coordinator: fix not calling build callbacks for dep-only builds
We cannot rely that the list of pending builds will contain all builds
that will happen when run_all is called. Some builds might only be
pulled as dependencies of others. Therefore, we need to keep a separate
future and manually fulfill it when the pending queue is emptied.

To handle that we must allow a callback to be specified, such that it
can be applied to builds that have not started when run_all was called.

Additionally, start handling the global build timeout, as Watcher now
expects a second cancellation whenever processes should be killed.